IVD-97 - Updating Cxense plugin to better follow plugin standard. Version 2.0.0: No more `Plugin` classes, no more deep folder structure, no more custom autoloading.
