+++ composer.json
@@ -1,15 +1,9 @@
 {
   "name": "benjaminmedia/wp-cxense",
-  "description": "Log in users from the WA platform",
+  "description": "This plugin integrates your site with cXense by adding meta tags and calling the cXense api",
   "type": "wordpress-plugin",
   "keywords": ["wordpress", "plugin"],
   "homepage": "https://github.com/BenjaminMedia/wp-cxense",
-  "repositories":[
-    {
-      "type": "composer",
-      "url": "http://wpackagist.org"
-    }
-  ],
   "authors": [
     {
       "name": "Bonnier Publications",
@@ -27,10 +21,10 @@
   },
   "autoload": {
     "psr-4": {
-      "Bonnier\\WP\\Cxense\\": "src/Bonnier/WP/Cxense/"
+      "Bonnier\\WP\\Cxense\\": "src"
     }
   },
   "require": {
-    "php": ">=5.6"
+    "php": ">=7.1"
   }
 }
