+++ src/Assets/Scripts.php
