+++ src/Exceptions/DocumentSearchMissingCount.php
