+++ src/Exceptions/DocumentSearchMissingSearch.php
