+++ src/Exceptions/DocumentSearchWrongFacet.php
