+++ src/Exceptions/DocumentSearchWrongFilter.php
