+++ src/Exceptions/DocumentSearchWrongSorting.php
