+++ src/Services/WidgetDocument.php
