+++ src/Settings/Partials/WidgetSettings.php
